rename 'types.ts' -> 'loadTypes.ts'
