+++ src/routes/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import type { LoadYear } from 'types/types';
+	import type { LoadYear } from 'types/loadTypes';
 	export let data: LoadYear; // data from database
 
 	import Year from './year/[year]/year.svelte';
